Remove LowPower dependency
